# [ci] speed up conda setup for some jobs