In #5648 , I saw that `conda create`-ing an environment in one shot was significantly faster in some situations than `conda  create`-ing an environment and then updating it with `conda install` or `conda update` (https://github.com/microsoft/LightGBM/issues/5647#issuecomment-1373148054).

That PR was for Windows CI jobs.

This PR proposes similar changes to some of the macOS and Linux jobs.

* moves `lint`, `if-else`, and `swig` job code up earlier in `.ci/test.sh`
   - *`swig` job doesn't need a conda environment, so time in that job spent waiting for `conda create` is wasted time*
   - *`lint` and `if-else` have very different `conda`-installed dependencies than the other jobs, so they can be totally self-contained and just `conda create` their own environment with everything they need*

### benefits of this change

I observed the following timing differences in these jobs on this PR, compared to `master`:

| job                                              | latest `master` | this PR   |
|:------------------------------|:---------------:|:--------:|
| (GHA) macOS if-else                | 06m33s             | 07m04s |
| (GHA) check-docs                    | 04m08s            | 03m30s  |
| (Azure) Linux inference            | 04m52s            | 04m36s |
| (Azure) Linux swig                    |  04m55s           | 04m02s |
| (Azure) Linux_latest inference|  04m07s            | 03m56s |
| (Azure) macOS swig                 | 07m42s             |05m05s  |

this PR:

* macOS GitHub Actions: https://github.com/microsoft/LightGBM/actions/runs/3870849816
* static_analysis GitHub Actions: https://github.com/microsoft/LightGBM/actions/runs/3870849804
* Azure: https://dev.azure.com/lightgbm-ci/lightgbm-ci/_build/results?buildId=14120&view=results

latest `master` build:

* macOS GitHub Actions: https://github.com/microsoft/LightGBM/actions/runs/3849602797
* static_analysis GitHub Actions: https://github.com/microsoft/LightGBM/actions/runs/3849602798
* Azure: https://dev.azure.com/lightgbm-ci/lightgbm-ci/_build/results?buildId=14109&view=results

### Notes for Reviewers

Given all the things that can change from run-to-run, it's hard to trust small differences in the timings based on 1 or even a few of runs. But I'm still pretty confident that these changes make these CI jobs at least *a small amount* faster, without reducing their stability.

If this PR is accepted, I'll propose another one after that consolidates these into a single `conda create` for the remaining tasks:

https://github.com/microsoft/LightGBM/blob/61e464bc02bd325672ce71daabfa57d569cf02bc/.ci/test.sh#L37

https://github.com/microsoft/LightGBM/blob/61e464bc02bd325672ce71daabfa57d569cf02bc/.ci/test.sh#L126-L140